GPT 3.5 Turbo Support, & Other Security Updates
Features Improvement:
GPT 3.5 Turbo Added by Default
/sc command for Source Code
Other Security enhancement

Version Update Information:

Updated the version from 1.0.2 to 2.0.0.
Modified the dependency "@adiwajshing/baileys" to use the custom GitHub repository "github:WhiskeySockets/Baileys" instead of the version 4.4.0.
Updated the version of "@adiwajshing/keyed-db" dependency to "^0.2.4".
Updated the version of "@hapi/boom" dependency to "^10.0.0".
Updated the version of "awesome-phonenumber" dependency to "^2.64.0".
Updated the version of "axios" dependency to "^1.3.4".
Updated the version of "chalk" dependency to "^4.1.2".
Updated the version of "chatgpt" dependency to "^4.1.1".
Updated the version of "crypto" dependency to "^1.0.1".
Updated the version of "figlet" dependency to "^1.5.2".
Updated the version of "fs" dependency to "0.0.1-security".
Updated the version of "lodash" dependency to "^4.17.21".
Updated the version of "openai" dependency to "^3.2.1".
Updated the version of "pino" dependency to "^7.0.5".
Updated the version of "qrcode-terminal" dependency to "^0.12.0".
Updated the version of "util" dependency to "^0.12.4".